:recycle: Rename package name for admin
